[gem] Update cucumber to 5.3.0
We will skip many versions from start.
It does not look like there are special needs to do for the upgrade in our case
