+++ features/support/0001_world.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+World(TimecopHelpers)
+World(PlanHelpers)
+World(DummyAttachments)
+
+World(JsHelpers)
+World(MessagesHelpers)
+World(BackendHelpers)
+World(AppsKeysHelpers)
+World(SiteAccount)
+World(ProviderNaming)
