+++ features/support/env.rb
@@ -1,50 +1,12 @@
-# frozen_string_literal: true
-
-if ENV['CI']
-  require 'simplecov'
-  require "simplecov_json_formatter"
-  require 'codecov'
-  formatters = [
-    SimpleCov::Formatter::SimpleFormatter,
-    SimpleCov::Formatter::JSONFormatter,
-    SimpleCov::Formatter::HTMLFormatter,
-    Codecov::SimpleCov::Formatter
-  ]
-  SimpleCov.start do
-    formatter SimpleCov::Formatter::MultiFormatter.new(formatters)
-  end
-end
-
 # IMPORTANT: This file is generated by cucumber-rails - edit at your own peril.
 # It is recommended to regenerate this file in the future when you upgrade to a
 # newer version of cucumber-rails. Consider adding your own code to a new file
 # instead of editing this one. Cucumber will automatically load all features/**/*.rb
 # files.
 
-require 'cucumber/formatter/unicode' # Remove this line if you don't want Cucumber Unicode support
-
-non_transactional = %w[
-  @backend
-  @emails
-  @stats
-  @audit
-  @commit-transactions
-].freeze
-
-transactional = non_transactional.map {|t| "not #{t}" }
-
-Before transactional.join(' or ') do
-  Cucumber::Rails::Database.before_js if Cucumber::Rails::Database.autorun_database_cleaner
-end
-
-Before non_transactional.join(' or ') do
-  Cucumber::Rails::Database.before_non_js if Cucumber::Rails::Database.autorun_database_cleaner
-end
-
 require 'cucumber/rails'
 
-# Lets you click links with onclick javascript handlers without using @culerity or @javascript
-# require 'cucumber/rails/capybara_javascript_emulation'
+# frozen_string_literal: true
 
 # Capybara defaults to CSS3 selectors rather than XPath.
 # If you'd prefer to use XPath, just uncomment this line and adjust any
@@ -71,10 +33,7 @@ ActionController::Base.allow_rescue = false
 # Remove/comment out the lines below if your app doesn't have a database.
 # For some databases (like MongoDB and CouchDB) you may need to use :truncation instead.
 begin
-  # Possible values are :truncation and :transaction
-  # The :transaction strategy is faster, but might give you threading problems.
-  # See https://github.com/cucumber/cucumber-rails/blob/master/features/choose_javascript_database_strategy.feature
-  Cucumber::Rails::Database.javascript_strategy = :truncation
+  DatabaseCleaner.strategy = :transaction
 rescue NameError
   raise "You need to add database_cleaner to your Gemfile (in the :test group) if you wish to use it."
 end
@@ -93,3 +52,8 @@ end
 #     DatabaseCleaner.strategy = :transaction
 #   end
 #
+
+# Possible values are :truncation and :transaction
+# The :transaction strategy is faster, but might give you threading problems.
+# See https://github.com/cucumber/cucumber-rails/blob/master/features/choose_javascript_database_strategy.feature
+Cucumber::Rails::Database.javascript_strategy = :truncation
